+++ src/index.d.ts
@@ -3,12 +3,13 @@ import { Body, Request, RequestInit } from 'node-fetch';
 
 type Opts<Data> = RequestInit & {
   bodyReader?: (b: Body) => Promise<Data>;
+  noSuspense?: boolean;
 };
 
 export type UseFetch = <Data>(input: string | Request, opts?: Opts<Data>) => {
-  error: Error | void;
-  loading: boolean;
-  data: Data | void;
+  error: Error | null;
+  loading?: boolean; // only used if noSuspense is true
+  data: Data | null;
   abort: () => void;
 };